remove entry details & start timesheet submit

+++ client/src/components/user/EntriesList.js
@@ -4,17 +4,13 @@ import moment from 'moment';
 import { getEntries, deleteEntry } from '../../actions/userActions';
 
 class EntriesList extends Component {
-    componentDidMount() {
-        const { getEntries, user } = this.props;
-        if (user) {
-            getEntries(user.user_id)
-        }
-    }
 
     handleDelete = async (entry) => {
         const { user, deleteEntry, getEntries } = this.props;
-        await deleteEntry(entry.entry_id);
-        getEntries(user.user_id);
+        if (window.confirm('Are you sure you want to delete?')) {
+            await deleteEntry(entry.entry_id);
+            getEntries(user.user_id);
+        }
     }
 
     render() {
@@ -28,7 +24,7 @@ class EntriesList extends Component {
                             <th>Project Name</th>
                             <th>Date</th>
                             <th>Hours Worked</th>
-                            <th>Details</th>
+                            <th>Delete</th>
                         </tr>
                     </thead>
 
@@ -40,7 +36,6 @@ class EntriesList extends Component {
                                         <td>{entry.project_name}</td>
                                         <td>{moment(entry.date).format('L')}</td>
                                         <td>{entry.hours_worked}</td>
-                                        <td>{entry.details}</td>
                                         <td><button className="btn red lighten-1" onClick={this.handleDelete.bind(this, entry)}>X</button></td>
                                     </tr>
                                 )) :
@@ -55,7 +50,6 @@ class EntriesList extends Component {
 
 const mapStateToProps = state => {
     return {
-        config: state.auth.config,
         user: state.auth.user,
         entries: state.user.entries
     }
